refactor: move bbs12381g2pub to pkg/crypto
