+++ pkg/crypto/api.go
@@ -7,7 +7,11 @@ SPDX-License-Identifier: Apache-2.0
 package crypto
 
 // package crypto contains the Crypto interface to be used by the framework.
-// it will be created via Options creation in pkg/framework/context.Provider
+// It will be created via Options creation in pkg/framework/context.Provider.
+// BBS+ signature scheme is not included in the main Crypto interface.
+// It is defined separately under the primitive sub-package including its implementation which should not be referenced
+// directly. It is accessible via the framework's KMS BBS+ keys and tinkcrypto's bbs package's Signer and Verifier
+// primitives or via webkms for remote KMS BBS+ signing.
 
 // Crypto interface provides all crypto operations needed in the Aries framework.
 type Crypto interface {
