+++ pkg/crypto/primitive/bbs12381g2pub/internal/kilic/bls12-381/hash_to_field_custom.go
