+++ pkg/crypto/primitive/bbs12381g2pub/internal/kilic/bls12-381/utils.go
