+++ pkg/crypto/primitive/bbs12381g2pub/proof_of_knowledge.go
