+++ pkg/crypto/primitive/bbs12381g2pub/signature_message.go
