+++ pkg/crypto/primitive/bbs12381g2pub/utils.go
