+++ pkg/crypto/tinkcrypto/primitive/bbs/bbs_verifier_key_manager_test.go
@@ -15,8 +15,8 @@ import (
 	"github.com/google/tink/go/subtle"
 	"github.com/stretchr/testify/require"
 
+	"github.com/hyperledger/aries-framework-go/pkg/crypto/primitive/bbs12381g2pub"
 	bbspb "github.com/hyperledger/aries-framework-go/pkg/crypto/tinkcrypto/primitive/proto/bbs_go_proto"
-	"github.com/hyperledger/aries-framework-go/pkg/doc/bbs/bbs12381g2pub"
 )
 
 func TestVerifierKeyManager_Primitive(t *testing.T) {
