+++ pkg/doc/signature/verifier/public_key_verifier.go
@@ -21,7 +21,7 @@ import (
 	"github.com/btcsuite/btcd/btcec"
 	gojose "github.com/square/go-jose/v3"
 
-	"github.com/hyperledger/aries-framework-go/pkg/doc/bbs/bbs12381g2pub"
+	"github.com/hyperledger/aries-framework-go/pkg/crypto/primitive/bbs12381g2pub"
 	"github.com/hyperledger/aries-framework-go/pkg/doc/jose"
 )
