+++ test/bbs/src/support_signer.go
@@ -11,7 +11,7 @@ package main
 import (
 	"strings"
 
-	bbs "github.com/hyperledger/aries-framework-go/pkg/doc/bbs/bbs12381g2pub"
+	bbs "github.com/hyperledger/aries-framework-go/pkg/crypto/primitive/bbs12381g2pub"
 )
 
 type bbsSigner struct {
